+++ app/Helper.php
@@ -1,5 +1,7 @@
 <?php
 
+use App\Models\Event;
+
 /**
  * Flash a message that can be used in layouts.alert
  *
@@ -12,4 +14,12 @@ function flashMessage($type, $title, $message)
     Session::flash('type', $type);
     Session::flash('title', $title);
     Session::flash('message', $message);
+}
+
+/**
+ * @return Event
+ */
+function nextEvent()
+{
+    return Event::where('endEvent', '>', now())->orderBy('startEvent', 'asc')->first();
 }