+++ app/Http/Controllers/Auth/RegisterController.php
@@ -2,12 +2,11 @@
 
 namespace App\Http\Controllers\Auth;
 
-use App\{
-    User, Http\Controllers\Controller
-};
-use Illuminate\{
-    Support\Facades\Hash, Support\Facades\Validator, Foundation\Auth\RegistersUsers
-};
+use App\User;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Foundation\Auth\RegistersUsers;
 
 class RegisterController extends Controller
 {