+++ app/Http/Resources/BookingsCollection.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\Http\Resources;
+
+use Illuminate\Http\Resources\Json\ResourceCollection;
+
+class BookingsCollection extends ResourceCollection
+{
+    /**
+     * Transform the resource collection into an array.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return array
+     */
+    public function toArray($request)
+    {
+        return [
+            'data' => BookingResource::collection($this->collection),
+            'meta' => ['bookings_count' => $this->collection->count()],
+        ];
+    }
+}