+++ app/Mail/EventNotification.php
@@ -2,7 +2,10 @@
 
 namespace App\Mail;
 
-use Illuminate\{Bus\Queueable, Contracts\Queue\ShouldQueue, Mail\Mailable, Queue\SerializesModels};
+use Illuminate\Bus\Queueable;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Mail\Mailable;
+use Illuminate\Queue\SerializesModels;
 
 class EventNotification extends Mailable implements ShouldQueue
 {