+++ database/migrations/2018_12_04_191327_add_airport_view_to_users_table.php
@@ -1,0 +1,33 @@
+<?php
+
+use App\Enums\AirportView;
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddAirportViewToUsersTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('users', function (Blueprint $table) {
+            $table->tinyInteger('airport_view')->after('isAdmin')->default(AirportView::NAME);
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('users', function (Blueprint $table) {
+            $table->dropColumn('airport_view');
+        });
+    }
+}