+++ resources/views/booking/overview.blade.php
@@ -15,7 +15,7 @@
                    class="btn btn-{{ url()->current() === route('bookings.event.index', $event) . '/arrivals' ? 'success' : 'primary' }}">Show
                     Arrivals</a>&nbsp;
             @endif
-            @if(Auth::check() && Auth::user()->isAdmin && $event->endBooking > now())
+            @if(Auth::check() && Auth::user()->isAdmin && $event->endBooking >= now())
                 <a href="{{ route('bookings.create',$event) }}" class="btn btn-primary"><i class="fa fa-plus"></i> Add
                     Booking</a>&nbsp;
                 <a href="{{ route('bookings.create',$event) }}/bulk" class="btn btn-primary"><i class="fa fa-plus"></i>
@@ -24,7 +24,7 @@
             @endif
         </p>
         @include('layouts.alert')
-        @if($event->startBooking < now() || Auth::check() && Auth::user()->isAdmin)
+        @if($event->startBooking <= now() || Auth::check() && Auth::user()->isAdmin)
             Flights available: {{ count($bookings) - count($bookings->where('status',\App\Enums\BookingStatus::BOOKED)) }} / {{ count($bookings) }}
             <table class="table table-hover">
                 <thead>
@@ -47,10 +47,30 @@
                     {{--Check if flight belongs to the logged in user--}}
                     <tr class="{{ Auth::check() && $booking->user_id == Auth::id() ? 'table-primary' : '' }}">
                         <td>
-                            <abbr title="{{ $booking->airportDep->name }}">{{ $booking->dep }}</abbr>
+                            @if(Auth::check())
+                                @if(Auth::user()->airport_view == \App\Enums\AirportView::NAME)
+                                    <abbr title="{{ $booking->dep }} | [{{ $booking->airportDep->iata }}]">{{ $booking->airportDep->name }}</abbr>
+                                @elseif(Auth::user()->airport_view == \App\Enums\AirportView::IATA)
+                                    <abbr title="{{ $booking->airportDep->name }} | [{{ $booking->dep }}]">{{ $booking->airportDep->iata }}</abbr>
+                                @else
+                                    <abbr title="{{ $booking->airportDep->name }} | [{{ $booking->airportDep->iata }}]">{{ $booking->dep }}</abbr>
+                                @endif
+                            @else
+                                <abbr title="{{ $booking->airportDep->name }} | [{{ $booking->airportDep->iata }}]">{{ $booking->dep }}</abbr>
+                            @endif
                         </td>
                         <td>
-                            <abbr title="{{ $booking->airportArr->name }}">{{ $booking->arr }}</abbr>
+                            @if(Auth::check())
+                                @if(Auth::user()->airport_view == \App\Enums\AirportView::NAME)
+                                    <abbr title="{{ $booking->arr }} | [{{ $booking->airportArr->iata }}]">{{ $booking->airportArr->name }}</abbr>
+                                @elseif(Auth::user()->airport_view == \App\Enums\AirportView::IATA)
+                                    <abbr title="{{ $booking->airportArr->name }} | [{{ $booking->arr }}]">{{ $booking->airportArr->iata }}</abbr>
+                                @else
+                                    <abbr title="{{ $booking->airportArr->name }} | [{{ $booking->airportArr->iata }}]">{{ $booking->arr }}</abbr>
+                                @endif
+                            @else
+                                <abbr title="{{ $booking->airportArr->name }} | [{{ $booking->airportArr->iata }}]">{{ $booking->arr }}</abbr>
+                            @endif
                         </td>
                         @if($booking->event->uses_times)
                             <td>
@@ -60,8 +80,8 @@
                                 {{ $booking->eta }}
                             </td>
                         @endif
-                        <td>{{ $booking->callsign }}</td>
-                        <td>{{ $booking->acType }}</td>
+                        <td class="{{ Auth::check() && Auth::user()->use_monospace_font ? 'text-monospace' : '' }}">{{ $booking->callsign }}</td>
+                        <td class="{{ Auth::check() && Auth::user()->use_monospace_font ? 'text-monospace' : '' }}">{{ $booking->acType }}</td>
                         <td>
                             {{--Check if booking has been booked--}}
                             @if($booking->status === \App\Enums\BookingStatus::BOOKED)
@@ -88,9 +108,9 @@
                             @else
                                 @if(Auth::check())
                                     {{--Check if user is logged in--}}
-                                    @if($booking->event->startBooking < now() && $booking->event->endBooking > now())
+                                    @if($booking->event->startBooking <= now() && $booking->event->endBooking >= now())
                                         {{--Check if user already has a booking--}}
-                                        @if(($booking->event->multiple_bookings_allowed) || (!$booking->event->multiple_bookings_allowed && !Auth::user()->booking()->where('event_id',$event->id)->first()))
+                                        @if(($booking->event->multiple_bookings_allowed) || (!$booking->event->multiple_bookings_allowed && !Auth::user()->bookings()->where('event_id',$event->id)->first()))
                                             {{--Check if user already has a booking, and only 1 is allowed--}}
                                             <a href="{{ route('bookings.edit', $booking) }}" class="btn btn-success">BOOK
                                                 NOW</a>
@@ -101,7 +121,7 @@
                                         <button class="btn btn-danger">Not available</button>
                                     @endif
                                 @else
-                                    <a href="{{ route('login') }}" class="btn btn-info">Click here to login</a>
+                                    <a href="{{ route('login', $booking) }}" class="btn btn-info">Click here to login</a>
                                 @endif
                             @endif
                         </td>