+++ resources/views/event/overview.blade.php
@@ -23,22 +23,20 @@
                 <tr>
                     @endif
                     <td>{{ $event->id }}</td>
-                    <td><a href="{{ route('bookings.event.index', $event) }}" title="Open Slot Table">{{ $event->name }}</a>
+                    <td><a href="{{ route('events.show', $event) }}">{{ $event->name }}</a>
                     </td>
                     <td>{{ $event->type->name }}</td>
                     <td>{{ $event->startEvent->format('d-m-Y') }}</td>
                     <td>{{ $event->startEvent->format('Hi') }}z</td>
                     <td>{{ $event->endEvent->format('Hi') }}z</td>
                     <td>
-                        <a href="{{ route('events.edit',$event->id) }}" role="button" class="btn btn-primary disabled"><i
+                        <a href="{{ route('events.edit',$event) }}" role="button" class="btn btn-primary"><i
                                     class="fa fa-edit"></i> Edit Event</a>&nbsp;
                         @if($event->endEvent > now())
                             <a href="{{ route('bookings.admin.importForm',$event) }}" class="btn btn-success"><i
                                         class="fa fa-edit"></i> Import data</a>&nbsp;
                             <a href="{{ route('bookings.create',$event) }}" class="btn btn-primary"><i
                                         class="fa fa-plus"></i> Add Timeslots</a>&nbsp;
-                            <a href="{{ route('bookings.admin.autoAssignForm',$event) }}" class="btn btn-primary">Auto
-                                Assign FL / Route</a>&nbsp;
                             @if($event->is_oceanic_event)
                                 <a href="{{ route('bookings.admin.autoAssignForm',$event) }}" class="btn btn-primary">
                                     Auto Assign FL / Route</a>&nbsp;
@@ -48,6 +46,13 @@
                                     class="fa fa-envelope"></i> Send E-mails (all persons)</a>&nbsp;
                         <a href="{{ route('bookings.export',$event) }}" class="btn btn-success"><i
                                     class="fa fa-edit"></i> Export data</a>&nbsp;
+                        @if($event->startEvent > now())
+                            <form action="{{ route('events.destroy', $event) }}" method="post">
+                                @csrf
+                                @method('DELETE')
+                                <button class="btn btn-danger">Delete Event</button>
+                            </form>
+                        @endif
                     </td>
                 </tr>
                 @empty